# fix: improve performance on reports log queries