--- a/superset/reports/models.py
+++ b/superset/reports/models.py
@@ -24,6 +24,7 @@ from sqlalchemy import (
     DateTime,
     Float,
     ForeignKey,
+    Index,
     Integer,
     String,
     Table,
@@ -184,6 +185,7 @@ class ReportRecipients(Model, AuditMixinNullable):
     """
 
     __tablename__ = "report_recipient"
+
     id = Column(Integer, primary_key=True)
     type = Column(String(50), nullable=False)
     recipient_config_json = Column(Text, default="{}")
@@ -194,6 +196,10 @@ class ReportRecipients(Model, AuditMixinNullable):
         ReportSchedule,
         backref=backref("recipients", cascade="all,delete,delete-orphan"),
         foreign_keys=[report_schedule_id],
+    )
+
+    __table_args__ = (
+        Index("ix_report_recipient_report_schedule_id", report_schedule_id),
     )
 
 
@@ -227,4 +233,10 @@ class ReportExecutionLog(Model):  # pylint: disable=too-few-public-methods
         ReportSchedule,
         backref=backref("logs", cascade="all,delete,delete-orphan"),
         foreign_keys=[report_schedule_id],
+    )
+
+    __table_args__ = (
+        Index("ix_report_execution_log_report_schedule_id", report_schedule_id),
+        Index("ix_report_execution_log_start_dttm", start_dttm),
+        Index("ix_report_execution_log_state", state),
     )