@dpgaspar 2024-01-09
> @dpgaspar do you have a sense how long this migration will take? Is there merit in adding a line item in `UPDATING.md` indicating that down time might be necessary?

Added details to the description and a line to UPDATING.md